utils/config_loader.py
import yaml

def load_config(config_path: str) -> dict:
    """Load a YAML configuration file and return it as a dictionary.

    Args:
        config_path: Path to the YAML configuration file.
    Returns:
        A dictionary containing the configuration.
    """
    with open(config_path, 'r') as file:
        config = yaml.safe_load(file)
    return config