pyproject.toml
[project]
name = "autovivqamodelbuilder"
version = "0.1.0"
description = "Graduation Project - Build a model with autoViVQA dataset"
authors = [
    {name = "tuong.nguyen",email = "richardnguyen0715@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <3.15"
dependencies = [
    "pyyaml (>=6.0.3,<7.0.0)",
    "kagglehub (==0.4.0)",
    "opencv-python (<5.0.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "torch (>=2.9.1,<3.0.0)",
    "torchvision (>=0.24.1,<0.25.0)",
    "transformers (>=4.57.2,<5.0.0)",
    "timm (>=1.0.22,<2.0.0)",
    "clip (>=0.2.0,<0.3.0)",
    "nltk (>=3.9.2,<4.0.0)",
    "psutil (>=7.1.3,<8.0.0)",
    "segment-anything (>=1.0,<2.0)",
    "underthesea (>=8.3.0,<9.0.0)",
    "vncorenlp (>=1.0.3,<2.0.0)",
    "chromadb (>=1.3.7,<2.0.0)",
    "faiss-cpu (>=1.13.1,<2.0.0)",
    "wandb (>=0.23.1,<0.24.0)",
    "tensorboard (>=2.20.0,<3.0.0)",
    "pynvml (>=13.0.1,<14.0.0)",
    "nvidia-ml-py (>=13.590.44,<14.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
